+++ lib/k8s/operation.ex
@@ -3,7 +3,10 @@ defmodule K8s.Operation do
 
   alias K8s.{Operation, Selector}
   alias K8s.Operation.Error
-  @derive {Jason.Encoder, except: [:path_params]}
+  @derive {Jason.Encoder, except: [:path_params, :header_params]}
+
+  @typedoc "Acceptable patch types"
+  @type patch_type :: :strategic_merge | :merge | :json_merge | :apply
 
   @allow_http_body [:put, :patch, :post]
   @selector :labelSelector
@@ -20,6 +23,13 @@ defmodule K8s.Operation do
     apply: :patch
   }
 
+  @patch_type_header_map %{
+    merge: ["Content-Type": "application/merge-patch+json"],
+    strategic_merge: ["Content-Type": "application/strategic-merge-patch+json"],
+    json_merge: ["Content-Type": "application/json-patch+json"],
+    apply: ["Content-Type": "application/apply-patch+yaml"]
+  }
+
   defstruct method: nil,
             verb: nil,
             api_version: nil,
@@ -27,7 +37,8 @@ defmodule K8s.Operation do
             data: nil,
             conn: nil,
             path_params: [],
-            query_params: []
+            query_params: [],
+            header_params: ["Content-Type": "application/json"]
 
   @typedoc "`K8s.Operation` name. May be an atom, string, or tuple of `{resource, subresource}`."
   @type name_t :: binary() | atom() | {binary(), binary()}
@@ -40,6 +51,7 @@ defmodule K8s.Operation do
   * `verb` - Kubernetes [REST API verb](https://kubernetes.io/docs/reference/access-authn-authz/authorization/#determine-the-request-verb) (`deletecollection`, `update`, `create`, `watch`, etc)
   * `path_params` - Parameters to interpolate into the Kubernetes REST URL
   * `query_params` - Query parameters. Merged w/ params provided to any `K8s.Client.Runner`. `K8s.Client.Runner` options win.
+  * `header_params` - Header parameters.
 
   `name` would be `deployments` in the case of a deployment, but may be `deployments/status` or `deployments/scale` for Status and Scale subresources.
 
@@ -54,7 +66,8 @@ defmodule K8s.Operation do
     api_version: "v1", # api version of the "Scale" kind
     name: "deployments/scale",
     data: %{"apiVersion" => "v1", "kind" => "Scale"}, # `data` is of kind "Scale"
-    path_params: [name: "nginx", namespace: "default"]
+    path_params: [name: "nginx", namespace: "default"],
+    header_params: ["Content-Type": "application/json"]
   }
   ```
 
@@ -67,7 +80,8 @@ defmodule K8s.Operation do
     api_version: "apps/v1", # api version of the "Deployment" kind
     name: "deployments/status",
     data: %{"apiVersion" => "apps/v1", "kind" => "Deployment"}, # `data` is of kind "Deployment"
-    path_params: [name: "nginx", namespace: "default"]
+    path_params: [name: "nginx", namespace: "default"],
+    header_params: ["Content-Type": "application/json"]
   }
   ```
   """
@@ -79,7 +93,8 @@ defmodule K8s.Operation do
           data: map() | nil,
           conn: K8s.Conn.t() | nil,
           path_params: keyword(),
-          query_params: keyword()
+          query_params: keyword(),
+          header_params: keyword()
         }
 
   @doc """
@@ -165,6 +180,7 @@ defmodule K8s.Operation do
   @spec build(atom, binary, name_t(), keyword(), map() | nil, keyword()) :: __MODULE__.t()
   def build(verb, api_version, name_or_kind, path_params, data \\ nil, opts \\ []) do
     http_method = @verb_map[verb] || verb
+    patch_type = Keyword.get(opts, :patch_type, :not_set)
 
     http_body =
       case http_method do
@@ -173,22 +189,44 @@ defmodule K8s.Operation do
       end
 
     query_params =
-      case verb do
-        :apply ->
+      cond do
+        verb === :apply || (verb === :patch && patch_type === :apply) ->
           [
             fieldManager: Keyword.get(opts, :field_manager, "elixir"),
             force: Keyword.get(opts, :force, true)
           ]
 
-        :connect ->
+        verb === :connect ->
           [stdin: true, stdout: true, stderr: true, tty: false]
           |> Keyword.merge(
             Keyword.take(opts, [:stdin, :stdout, :stderr, :tty, :command, :container])
           )
 
-        _ ->
+        true ->
           []
       end
+
+    header_params =
+      case {verb, patch_type} do
+        {:patch, merge_patch_types} when merge_patch_types in [:merge, :not_set] ->
+          @patch_type_header_map[:merge]
+
+        {:patch, :strategic_merge} ->
+          @patch_type_header_map[:strategic_merge]
+
+        {:patch, :json_merge} ->
+          @patch_type_header_map[:json_merge]
+
+        {:patch, :apply} ->
+          @patch_type_header_map[:apply]
+
+        {:apply, apply_patch_types} when apply_patch_types in [:apply, :not_set] ->
+          @patch_type_header_map[:apply]
+
+        _ ->
+          ["Content-Type": "application/json"]
+      end
+      |> Keyword.merge(Keyword.get(opts, :header_params, []))
 
     %__MODULE__{
       method: http_method,
@@ -197,7 +235,8 @@ defmodule K8s.Operation do
       api_version: api_version,
       name: name_or_kind,
       path_params: path_params,
-      query_params: query_params
+      query_params: query_params,
+      header_params: header_params
     }
   end